BUGFIX - Change name up upvote in props to avoid name collision
There is a function named upVote in actions/index.js that gets
included in the Cover's props via the connect function at the bottom of
Cover/index.js

This was causing the onClick event in Cover/index.js to call the one in
actions rather than the one passed to it by ComicList.

NOTE: this is a pretty quick and dirty fix.  A better approach would be
to split apart the actions in actions/index.js to their own files and
include only the ones that are needed, rather than pollute the whole
namespace